Sửa kiểu dữ liệu

+++ src/red/team/boss/list_boss/Mabu12h/BuiBui.java
@@ -22,8 +22,8 @@ public class BuiBui extends Boss {
 
     @Override
     public void reward(Player plKill) {
-        byte randomDo = (byte) new Random().nextInt(Manager.itemIds_TL.length - 1);
-//        byte randomNR = (byte) new Random().nextInt(Manager.itemIds_NR_SB.length);
+        int randomDo = (int) new Random().nextInt(Manager.itemIds_TL.length - 1);
+//        int randomNR = (byte) new Random().nextInt(Manager.itemIds_NR_SB.length);
         int[] itemDos = new int[]{555,556,557,558,559,560,561,562,563,564,565,566,567};
         int randomc12 = new Random().nextInt(itemDos.length);
         if (Util.isTrue(BossManager.ratioReward, 100)) {
@@ -41,7 +41,7 @@ public class BuiBui extends Boss {
             
         } else {
             Service.gI().dropItemMap(this.zone,
-                    new ItemMap(zone, (byte) 16, 1, this.location.x, this.location.y, plKill.id));
+                    new ItemMap(zone, (int) 16, 1, this.location.x, this.location.y, plKill.id));
         }
         plKill.pointBoss += 0;
         plKill.fightMabu.changePoint(10);
