+++ src/red/team/boss/list_boss/Mabu12h/Yacon.java
@@ -21,7 +21,7 @@ public class Yacon extends Boss {
 
     @Override
     public void reward(Player plKill) {
-        byte randomDo = (byte) new Random().nextInt(Manager.itemIds_TL.length - 1);
+        short randomDo = (short) new Random().nextInt(Manager.itemIds_TL.length - 1);
 //        byte randomNR = (byte) new Random().nextInt(Manager.itemIds_NR_SB.length);
         int[] itemDos = new int[]{555,556,557,558,559,560,561,562,563,564,565,566,567};
         int randomc12 = new Random().nextInt(itemDos.length);
